Renames + re-namespaces + split out partials
- `CardHand` renamed to `HandOfCards, and similar classes.
- `HandCard` nested partial class split to separate file
(`HandOfCards.HandCard`).
- `Cards` namespace renamed to `Card`.
- Hand-related classes moved to the `Hand` namespace.
- `CardRow` renamed to `CardFlow`, renamespaced under `Components`.
