+++ osu.Game.Tests/Visual/RankedPlay/TestSceneRankedPlayScreen.cs
@@ -11,7 +11,7 @@ using osu.Game.Online.API.Requests.Responses;
 using osu.Game.Online.Multiplayer.MatchTypes.RankedPlay;
 using osu.Game.Online.Rooms;
 using osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay;
-using osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards;
+using osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Hand;
 using osuTK.Input;
 
 namespace osu.Game.Tests.Visual.RankedPlay
@@ -50,7 +50,7 @@ namespace osu.Game.Tests.Visual.RankedPlay
                 int i2 = i;
                 AddStep($"click card {i2}", () =>
                 {
-                    InputManager.MoveMouseTo(this.ChildrenOfType<PlayerCardHand.PlayerHandCard>().ElementAt(i2));
+                    InputManager.MoveMouseTo(this.ChildrenOfType<PlayerHandOfCards.PlayerHandCard>().ElementAt(i2));
                     InputManager.Click(MouseButton.Left);
                 });
             }
@@ -123,14 +123,14 @@ namespace osu.Game.Tests.Visual.RankedPlay
         public void TestPlayStage()
         {
             AddStep("set play phase", () => MultiplayerClient.RankedPlayChangeStage(RankedPlayStage.CardPlay, state => state.ActiveUserId = API.LocalUser.Value.OnlineID).WaitSafely());
-            AddUntilStep("wait until cards are present", () => this.ChildrenOfType<PlayerCardHand.PlayerHandCard>().Count() == 5);
+            AddUntilStep("wait until cards are present", () => this.ChildrenOfType<PlayerHandOfCards.PlayerHandCard>().Count() == 5);
 
             for (int i = 0; i < 3; i++)
             {
                 int i2 = i;
                 AddStep($"click card {i2}", () =>
                 {
-                    InputManager.MoveMouseTo(this.ChildrenOfType<PlayerCardHand.PlayerHandCard>().ElementAt(i2));
+                    InputManager.MoveMouseTo(this.ChildrenOfType<PlayerHandOfCards.PlayerHandCard>().ElementAt(i2));
                     InputManager.Click(MouseButton.Left);
                 });
             }
@@ -140,7 +140,7 @@ namespace osu.Game.Tests.Visual.RankedPlay
             AddStep("click play button", () =>
             {
                 var button = screen
-                             .ChildrenOfType<PlayerCardHand.PlayerHandCard>()
+                             .ChildrenOfType<PlayerHandOfCards.PlayerHandCard>()
                              .First(it => it.Selected)
                              .ChildrenOfType<ShearedButton>()
                              .First();
