+++ osu.Game/Screens/OnlinePlay/Matchmaking/RankedPlay/Card/RankedPlayCardExtensions.cs
@@ -6,7 +6,7 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Primitives;
 using osuTK;
 
-namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
+namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Card
 {
     public static class RankedPlayCardExtensions
     {
