+++ osu.Game/Screens/OnlinePlay/Matchmaking/RankedPlay/Hand/HandOfCards.cs
@@ -6,22 +6,22 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
 using osu.Framework.Caching;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Primitives;
 using osu.Game.Online.Multiplayer.MatchTypes.RankedPlay;
 using osu.Game.Online.RankedPlay;
+using osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Card;
 using osuTK;
 
-namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
+namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Hand
 {
     /// <summary>
     /// Drawable that layouts cards as if held in a player's hands.
     /// </summary>
     [Cached]
-    public abstract partial class CardHand : CompositeDrawable
+    public abstract partial class HandOfCards : CompositeDrawable
     {
         private const float hover_scale = 1.2f;
 
@@ -35,7 +35,7 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
 
         /// <summary>
         /// If true, card layout will be flipped on both axes for a card hand placed at the top edge of the screen, while keeping the cards upright.
-        /// Used for <see cref="OpponentCardHand"/>.
+        /// Used for <see cref="OpponentHandOfCards"/>.
         /// </summary>
         protected virtual bool Flipped => false;
 
@@ -43,7 +43,7 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
 
         private readonly Dictionary<RankedPlayCardItem, HandCard> cardLookup = new Dictionary<RankedPlayCardItem, HandCard>();
 
-        protected CardHand()
+        protected HandOfCards()
         {
             AddInternal(cardContainer = new Container<HandCard>
             {
@@ -251,101 +251,5 @@ namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
         }
 
         #endregion
-
-        public partial class HandCard : CompositeDrawable
-        {
-            public float LayoutWidth => DrawWidth * (State.Hovered ? hover_scale : 1);
-
-            private readonly Bindable<RankedPlayCardState> state = new Bindable<RankedPlayCardState>();
-
-            public RankedPlayCardState State
-            {
-                get => state.Value;
-                set => state.Value = value;
-            }
-
-            public bool Selected
-            {
-                get => State.Selected;
-                set => State = State with { Selected = value };
-            }
-
-            public bool CardHovered
-            {
-                get => State.Hovered;
-                set => State = State with { Hovered = value };
-            }
-
-            public bool CardPressed
-            {
-                get => State.Pressed;
-                set => State = State with { Pressed = value };
-            }
-
-            [Resolved]
-            private CardHand cardHand { get; set; } = null!;
-
-            public readonly RankedPlayCard Card;
-
-            public RankedPlayCardWithPlaylistItem Item => Card.Item;
-
-            public HandCard(RankedPlayCard card)
-            {
-                Size = card.DrawSize;
-
-                card.Anchor = Anchor.Centre;
-                card.Origin = Anchor.Centre;
-                card.Position = Vector2.Zero;
-                card.Rotation = 0;
-                card.Scale = Vector2.One;
-
-                AddInternal(Card = card);
-
-                Anchor = Anchor.BottomCentre;
-                Origin = Anchor.BottomCentre;
-            }
-
-            protected override void LoadComplete()
-            {
-                base.LoadComplete();
-
-                state.BindValueChanged(OnStateChanged, true);
-            }
-
-            protected virtual void OnStateChanged(ValueChangedEvent<RankedPlayCardState> state)
-            {
-                cardHand.OnCardStateChanged(this, state.NewValue);
-
-                Card.ShowSelectionOutline = state.NewValue.Selected;
-
-                switch (state.NewValue.Pressed, state.OldValue.Pressed)
-                {
-                    case (true, false):
-                        Card.ScaleTo(0.95f, 300, Easing.OutExpo);
-                        break;
-
-                    case (false, true):
-                        Card.ScaleTo(1f, 400, Easing.OutElasticHalf);
-                        break;
-                }
-            }
-
-            public RankedPlayCard Detach()
-            {
-                Card.ShowSelectionOutline = false;
-                Card.Elevation = 0;
-
-                RemoveInternal(Card, false);
-
-                return Card;
-            }
-
-            protected override void Update()
-            {
-                base.Update();
-
-                Card.Elevation = float.Lerp(CardHovered ? 1 : 0, Card.Elevation, (float)Math.Exp(-0.03f * Time.Elapsed));
-            }
-        }
     }
 }
