+++ osu.Game/Screens/OnlinePlay/Matchmaking/RankedPlay/Hand/HandSelectionMode.cs
@@ -1,9 +1,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Cards
+namespace osu.Game.Screens.OnlinePlay.Matchmaking.RankedPlay.Hand
 {
-    public enum CardSelectionMode
+    public enum HandSelectionMode
     {
         Disabled,
         Single,
